Fixed check.m4 so that --without-check is a valid option to disable check. Bumped version number to 0.8.3.
git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@135 64e312b2-a51f-0410-8e61-82d0ca0eb02a

+++ check/check.m4
@@ -7,26 +7,30 @@ AC_DEFUN(AM_PATH_CHECK,
   AC_ARG_WITH(check,
   [  --with-check=PATH       prefix where check is installed [default=auto]])
  
-  if test "x$with_check" != x; then
-    CHECK_CFLAGS="-I$with_check/include"
-    CHECK_LIBS="-L$with_check/lib -lcheck"
-  else
-    CHECK_CFLAGS=""
-    CHECK_LIBS="-lcheck"
-  fi
-
   min_check_version=ifelse([$1], ,0.8.2,$1)
 
   AC_MSG_CHECKING(for check - version >= $min_check_version)
 
-  ac_save_CFLAGS="$CFLAGS"
-  ac_save_LIBS="$LIBS"
+  if test x$with_check = xno; then
+    AC_MSG_RESULT(disabled)
+    ifelse([$3], , AC_MSG_ERROR([disabling check is not supported]), [$3])
+  else
+    if test "x$with_check" != x; then
+      CHECK_CFLAGS="-I$with_check/include"
+      CHECK_LIBS="-L$with_check/lib -lcheck"
+    else
+      CHECK_CFLAGS=""
+      CHECK_LIBS="-lcheck"
+    fi
 
-  CFLAGS="$CFLAGS $CHECK_CFLAGS"
-  LIBS="$CHECK_LIBS $LIBS"
+    ac_save_CFLAGS="$CFLAGS"
+    ac_save_LIBS="$LIBS"
 
-  rm -f conf.check-test
-  AC_TRY_RUN([
+    CFLAGS="$CFLAGS $CHECK_CFLAGS"
+    LIBS="$CHECK_LIBS $LIBS"
+
+    rm -f conf.check-test
+    AC_TRY_RUN([
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -79,21 +83,21 @@ int main ()
 }
 ],, no_check=yes, [echo $ac_n "cross compiling; assumed OK... $ac_c"])
 
-  CFLAGS="$ac_save_CFLAGS"
-  LIBS="$ac_save_LIBS"
+    CFLAGS="$ac_save_CFLAGS"
+    LIBS="$ac_save_LIBS"
 
-  if test "x$no_check" = x ; then
-    AC_MSG_RESULT(yes)
-    ifelse([$2], , :, [$2])
-  else
-    AC_MSG_RESULT(no)
-    if test -f conf.check-test ; then
-      :
+    if test "x$no_check" = x ; then
+      AC_MSG_RESULT(yes)
+      ifelse([$2], , :, [$2])
     else
-      echo "*** Could not run check test program, checking why..."
-      CFLAGS="$CFLAGS $CHECK_CFLAGS"
-      LIBS="$CHECK_LIBS $LIBS"
-      AC_TRY_LINK([
+      AC_MSG_RESULT(no)
+      if test -f conf.check-test ; then
+        :
+      else
+        echo "*** Could not run check test program, checking why..."
+        CFLAGS="$CFLAGS $CHECK_CFLAGS"
+        LIBS="$CHECK_LIBS $LIBS"
+        AC_TRY_LINK([
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -109,19 +113,21 @@ int main ()
       [ echo "*** The test program failed to compile or link. See the file config.log for"
         echo "*** the exact error that occured." ])
       
-      CFLAGS="$ac_save_CFLAGS"
-      LIBS="$ac_save_LIBS"
+        CFLAGS="$ac_save_CFLAGS"
+        LIBS="$ac_save_LIBS"
+      fi
+
+      CHECK_CFLAGS=""
+      CHECK_LIBS=""
+
+      rm -f conf.check-test
+      ifelse([$3], , AC_MSG_ERROR([check not found]), [$3])
     fi
 
-    CHECK_CFLAGS=""
-    CHECK_LIBS=""
+    AC_SUBST(CHECK_CFLAGS)
+    AC_SUBST(CHECK_LIBS)
 
     rm -f conf.check-test
-    ifelse([$3], , AC_MSG_ERROR([check not found]), [$3])
+
   fi
-
-  AC_SUBST(CHECK_CFLAGS)
-  AC_SUBST(CHECK_LIBS)
-
-  rm -f conf.check-test
 ])
